+++ packages/jest-config/src/getCacheDirectory.ts
@@ -3,13 +3,10 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-const path = require('path');
-const os = require('os');
-
+import path from 'path';
+import os from 'os';
 import {sync as realpath} from 'realpath-native';
 
 const getCacheDirectory = () => {