+++ packages/jest-config/src/normalize.ts
@@ -3,32 +3,22 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-import type {Argv} from 'types/Argv';
-import type {
-  InitialOptions,
-  DefaultOptions,
-  ReporterConfig,
-  GlobalConfig,
-  Path,
-  ProjectConfig,
-} from 'types/Config';
-
 import crypto from 'crypto';
-import glob from 'glob';
 import path from 'path';
+import glob from 'glob';
+import {Config} from '@jest/types';
+// @ts-ignore: Not migrated to TS
 import {ValidationError, validate} from 'jest-validate';
-import validatePattern from './validatePattern';
 import {clearLine, replacePathSepForGlob} from 'jest-util';
 import chalk from 'chalk';
-import getMaxWorkers from './getMaxWorkers';
 import micromatch from 'micromatch';
 import {sync as realpath} from 'realpath-native';
 import Resolver from 'jest-resolve';
 import {replacePathSepForRegex} from 'jest-regex-util';
+import validatePattern from './validatePattern';
+import getMaxWorkers from './getMaxWorkers';
 import {
   BULLET,
   DOCUMENTATION_NOTE,
@@ -46,18 +36,22 @@ import DEFAULT_CONFIG from './Defaults';
 import DEPRECATED_CONFIG from './Deprecated';
 import setFromArgv from './setFromArgv';
 import VALID_CONFIG from './ValidConfig';
-
 const ERROR = `${BULLET}Validation Error`;
 const PRESET_EXTENSIONS = ['.json', '.js'];
 const PRESET_NAME = 'jest-preset';
 
-const createConfigError = message =>
+type AllOptions = Config.ProjectConfig & Config.GlobalConfig;
+
+const createConfigError = (message: string) =>
   new ValidationError(ERROR, message, DOCUMENTATION_NOTE);
 
 const mergeOptionWithPreset = (
-  options: InitialOptions,
-  preset: InitialOptions,
-  optionName: string,
+  options: Config.InitialOptions,
+  preset: Config.InitialOptions,
+  optionName: keyof Pick<
+    Config.InitialOptions,
+    'moduleNameMapper' | 'transform'
+  >,
 ) => {
   if (options[optionName] && preset[optionName]) {
     options[optionName] = {
@@ -69,10 +63,10 @@ const mergeOptionWithPreset = (
 };
 
 const setupPreset = (
-  options: InitialOptions,
+  options: Config.InitialOptions,
   optionsPreset: string,
-): InitialOptions => {
-  let preset;
+): Config.InitialOptions => {
+  let preset: Config.InitialOptions;
   const presetPath = replaceRootDirInPath(options.rootDir, optionsPreset);
   const presetModule = Resolver.findNodeModule(
     presetPath.startsWith('.')
@@ -92,8 +86,8 @@ const setupPreset = (
       }
     } catch (e) {}
 
-    // $FlowFixMe
-    preset = (require(presetModule): InitialOptions);
+    // @ts-ignore: `presetModule` can be null?
+    preset = require(presetModule);
   } catch (error) {
     if (error instanceof SyntaxError || error instanceof TypeError) {
       throw createConfigError(
@@ -132,7 +126,7 @@ const setupPreset = (
   return {...preset, ...options};
 };
 
-const setupBabelJest = (options: InitialOptions) => {
+const setupBabelJest = (options: Config.InitialOptions) => {
   const transform = options.transform;
   let babelJest;
   if (transform) {
@@ -175,12 +169,16 @@ const setupBabelJest = (options: InitialOptions) => {
 };
 
 const normalizeCollectCoverageOnlyFrom = (
-  options: InitialOptions,
-  key: string,
+  options: Config.InitialOptions &
+    Required<Pick<Config.InitialOptions, 'collectCoverageOnlyFrom'>>,
+  key: keyof Pick<Config.InitialOptions, 'collectCoverageOnlyFrom'>,
 ) => {
-  const collectCoverageOnlyFrom = Array.isArray(options[key])
-    ? options[key] // passed from argv
-    : Object.keys(options[key]); // passed from options
+  const initialCollectCoverageFrom = options[key];
+  const collectCoverageOnlyFrom: Config.Glob[] = Array.isArray(
+    initialCollectCoverageFrom,
+  )
+    ? initialCollectCoverageFrom // passed from argv
+    : Object.keys(initialCollectCoverageFrom); // passed from options
   return collectCoverageOnlyFrom.reduce((map, filePath) => {
     filePath = path.resolve(
       options.rootDir,
@@ -191,22 +189,27 @@ const normalizeCollectCoverageOnlyFrom = (
   }, Object.create(null));
 };
 
-const normalizeCollectCoverageFrom = (options: InitialOptions, key: string) => {
-  let value;
-  if (!options[key]) {
+const normalizeCollectCoverageFrom = (
+  options: Config.InitialOptions &
+    Required<Pick<Config.InitialOptions, 'collectCoverageFrom'>>,
+  key: keyof Pick<Config.InitialOptions, 'collectCoverageFrom'>,
+) => {
+  const initialCollectCoverageFrom = options[key];
+  let value: Config.Glob[] | undefined;
+  if (!initialCollectCoverageFrom) {
     value = [];
   }
 
-  if (!Array.isArray(options[key])) {
+  if (!Array.isArray(initialCollectCoverageFrom)) {
     try {
-      value = JSON.parse(options[key]);
+      value = JSON.parse(initialCollectCoverageFrom);
     } catch (e) {}
 
     if (options[key] && !Array.isArray(value)) {
-      value = [options[key]];
+      value = [initialCollectCoverageFrom];
     }
   } else {
-    value = options[key];
+    value = initialCollectCoverageFrom;
   }
 
   if (value) {
@@ -219,8 +222,16 @@ const normalizeCollectCoverageFrom = (options: InitialOptions, key: string) => {
 };
 
 const normalizeUnmockedModulePathPatterns = (
-  options: InitialOptions,
-  key: string,
+  options: Config.InitialOptions,
+  key: keyof Pick<
+    Config.InitialOptions,
+    | 'coveragePathIgnorePatterns'
+    | 'modulePathIgnorePatterns'
+    | 'testPathIgnorePatterns'
+    | 'transformIgnorePatterns'
+    | 'watchPathIgnorePatterns'
+    | 'unmockedModulePathPatterns'
+  >,
 ) =>
   // _replaceRootDirTags is specifically well-suited for substituting
   // <rootDir> in paths (it deals with properly interpreting relative path
@@ -228,11 +239,13 @@ const normalizeUnmockedModulePathPatterns = (
   //
   // For patterns, direct global substitution is far more ideal, so we
   // special case substitutions for patterns here.
-  options[key].map(pattern =>
+  options[key]!.map(pattern =>
     replacePathSepForRegex(pattern.replace(/<rootDir>/g, options.rootDir)),
   );
 
-const normalizePreprocessor = (options: InitialOptions): InitialOptions => {
+const normalizePreprocessor = (
+  options: Config.InitialOptions,
+): Config.InitialOptions => {
   if (options.scriptPreprocessor && options.transform) {
     throw createConfigError(
       `  Options: ${chalk.bold('scriptPreprocessor')} and ${chalk.bold(
@@ -269,10 +282,10 @@ const normalizePreprocessor = (options: InitialOptions): InitialOptions => {
 };
 
 const normalizeMissingOptions = (
-  options: InitialOptions,
-  configPath: ?Path,
+  options: Config.InitialOptions,
+  configPath: Config.Path | null | undefined,
   projectIndex: number,
-): InitialOptions => {
+): Config.InitialOptions => {
   if (!options.name) {
     options.name = crypto
       .createHash('md5')
@@ -290,7 +303,9 @@ const normalizeMissingOptions = (
   return options;
 };
 
-const normalizeRootDir = (options: InitialOptions): InitialOptions => {
+const normalizeRootDir = (
+  options: Config.InitialOptions,
+): Config.InitialOptions => {
   // Assert that there *is* a rootDir
   if (!options.hasOwnProperty('rootDir')) {
     throw createConfigError(
@@ -309,7 +324,7 @@ const normalizeRootDir = (options: InitialOptions): InitialOptions => {
   return options;
 };
 
-const normalizeReporters = (options: InitialOptions, basedir) => {
+const normalizeReporters = (options: Config.InitialOptions) => {
   const reporters = options.reporters;
   if (!reporters || !Array.isArray(reporters)) {
     return options;
@@ -317,7 +332,7 @@ const normalizeReporters = (options: InitialOptions, basedir) => {
 
   validateReporters(reporters);
   options.reporters = reporters.map(reporterConfig => {
-    const normalizedReporterConfig: ReporterConfig =
+    const normalizedReporterConfig: Config.ReporterConfig =
       typeof reporterConfig === 'string'
         ? // if reporter config is a string, we wrap it in an array
           // and pass an empty object for options argument, to normalize
@@ -348,7 +363,7 @@ const normalizeReporters = (options: InitialOptions, basedir) => {
   return options;
 };
 
-const buildTestPathPattern = (argv: Argv): string => {
+const buildTestPathPattern = (argv: Config.Argv): string => {
   const patterns = [];
 
   if (argv._) {
@@ -386,11 +401,14 @@ const showTestPathPatternError = (testPathPattern: string) => {
 };
 
 export default function normalize(
-  options: InitialOptions,
-  argv: Argv,
-  configPath: ?Path,
-  projectIndex?: number = Infinity,
-) {
+  options: Config.InitialOptions,
+  argv: Config.Argv,
+  configPath?: Config.Path | null,
+  projectIndex: number = Infinity,
+): {
+  hasDeprecationWarnings: boolean;
+  options: AllOptions;
+} {
   const {hasDeprecationWarnings} = validate(options, {
     comment: DOCUMENTATION_NOTE,
     deprecatedConfig: DEPRECATED_CONFIG,
@@ -429,14 +447,12 @@ export default function normalize(
     options.setupTestFrameworkScriptFile &&
     options.setupFilesAfterEnv.length > 0
   ) {
-    throw createConfigError(
-      `  Options: ${chalk.bold(
-        'setupTestFrameworkScriptFile',
-      )} and ${chalk.bold('setupFilesAfterEnv')} cannot be used together.
+    throw createConfigError(`  Options: ${chalk.bold(
+      'setupTestFrameworkScriptFile',
+    )} and ${chalk.bold('setupFilesAfterEnv')} cannot be used together.
   Please change your configuration to only use ${chalk.bold(
     'setupFilesAfterEnv',
-  )}.`,
-    );
+  )}.`);
   }
 
   if (options.setupTestFrameworkScriptFile) {
@@ -452,6 +468,7 @@ export default function normalize(
     options.roots = options.testPathDirs;
     delete options.testPathDirs;
   }
+
   if (!options.roots) {
     options.roots = [options.rootDir];
   }
@@ -465,10 +482,10 @@ export default function normalize(
   }
 
   setupBabelJest(options);
-
-  const newOptions: $Shape<DefaultOptions & ProjectConfig & GlobalConfig> = {
+  // TODO: Type this properly
+  const newOptions = ({
     ...DEFAULT_CONFIG,
-  };
+  } as unknown) as AllOptions;
 
   try {
     // try to resolve windows short paths, ignoring errors (permission errors, mostly)
@@ -485,51 +502,68 @@ export default function normalize(
     });
   }
 
-  Object.keys(options).reduce((newOptions, key: $Keys<InitialOptions>) => {
+  const optionKeys = Object.keys(options) as Array<keyof Config.InitialOptions>;
+
+  optionKeys.reduce((newOptions, key: keyof Config.InitialOptions) => {
     // The resolver has been resolved separately; skip it
     if (key === 'resolver') {
       return newOptions;
     }
+
+    // This is cheating, because it claims that all keys of InitialOptions are Required.
+    // We only really know it's Required for oldOptions[key], not for oldOptions.someOtherKey,
+    // so oldOptions[key] is the only way it should be used.
+    const oldOptions = options as Config.InitialOptions &
+      Required<Pick<Config.InitialOptions, typeof key>>;
     let value;
     switch (key) {
       case 'collectCoverageOnlyFrom':
-        value = normalizeCollectCoverageOnlyFrom(options, key);
+        value = normalizeCollectCoverageOnlyFrom(oldOptions, key);
         break;
       case 'setupFiles':
       case 'setupFilesAfterEnv':
       case 'snapshotSerializers':
-        value =
-          options[key] &&
-          options[key].map(filePath =>
-            resolve(newOptions.resolver, {
-              filePath,
-              key,
-              rootDir: options.rootDir,
-            }),
-          );
+        {
+          const option = oldOptions[key];
+          value =
+            option &&
+            option.map(filePath =>
+              resolve(newOptions.resolver, {
+                filePath,
+                key,
+                rootDir: options.rootDir,
+              }),
+            );
+        }
         break;
       case 'modulePaths':
       case 'roots':
-        value =
-          options[key] &&
-          options[key].map(filePath =>
-            path.resolve(
-              options.rootDir,
-              replaceRootDirInPath(options.rootDir, filePath),
-            ),
-          );
+        {
+          const option = oldOptions[key];
+          value =
+            option &&
+            option.map(filePath =>
+              path.resolve(
+                options.rootDir,
+                replaceRootDirInPath(options.rootDir, filePath),
+              ),
+            );
+        }
         break;
       case 'collectCoverageFrom':
-        value = normalizeCollectCoverageFrom(options, key);
+        value = normalizeCollectCoverageFrom(oldOptions, key);
         break;
       case 'cacheDirectory':
       case 'coverageDirectory':
-        value =
-          options[key] &&
-          path.resolve(
-            options.rootDir,
-            replaceRootDirInPath(options.rootDir, options[key]),
-          );
+        {
+          const option = oldOptions[key];
+          value =
+            option &&
+            path.resolve(
+              options.rootDir,
+              replaceRootDirInPath(options.rootDir, option),
+            );
+        }
         break;
       case 'dependencyExtractor':
       case 'globalSetup':
@@ -539,37 +573,48 @@ export default function normalize(
       case 'testResultsProcessor':
       case 'testRunner':
       case 'filter':
-        value =
-          options[key] &&
-          resolve(newOptions.resolver, {
-            filePath: options[key],
-            key,
-            rootDir: options.rootDir,
-          });
+        {
+          const option = oldOptions[key];
+          value =
+            option &&
+            resolve(newOptions.resolver, {
+              filePath: option,
+              key,
+              rootDir: options.rootDir,
+            });
+        }
         break;
       case 'runner':
-        value =
-          options[key] &&
-          getRunner(newOptions.resolver, {
-            filePath: options[key],
-            rootDir: options.rootDir,
-          });
+        {
+          const option = oldOptions[key];
+          value =
+            option &&
+            getRunner(newOptions.resolver, {
+              filePath: option,
+              rootDir: options.rootDir,
+            });
+        }
         break;
       case 'prettierPath':
-        // We only want this to throw if "prettierPath" is explicitly passed
-        // from config or CLI, and the requested path isn't found. Otherwise we
-        // set it to null and throw an error lazily when it is used.
-        value =
-          options[key] &&
-          resolve(newOptions.resolver, {
-            filePath: options[key],
-            key,
-            optional: options[key] === DEFAULT_CONFIG[key],
-            rootDir: options.rootDir,
-          });
+        {
+          // We only want this to throw if "prettierPath" is explicitly passed
+          // from config or CLI, and the requested path isn't found. Otherwise we
+          // set it to null and throw an error lazily when it is used.
+
+          const option = oldOptions[key];
+
+          value =
+            option &&
+            resolve(newOptions.resolver, {
+              filePath: option,
+              key,
+              optional: option === DEFAULT_CONFIG[key],
+              rootDir: options.rootDir,
+            });
+        }
         break;
       case 'moduleNameMapper':
-        const moduleNameMapper = options[key];
+        const moduleNameMapper = oldOptions[key];
         value =
           moduleNameMapper &&
           Object.keys(moduleNameMapper).map(regex => {
@@ -578,7 +623,7 @@ export default function normalize(
           });
         break;
       case 'transform':
-        const transform = options[key];
+        const transform = oldOptions[key];
         value =
           transform &&
           Object.keys(transform).map(regex => [
@@ -596,12 +641,12 @@ export default function normalize(
       case 'transformIgnorePatterns':
       case 'watchPathIgnorePatterns':
       case 'unmockedModulePathPatterns':
-        value = normalizeUnmockedModulePathPatterns(options, key);
+        value = normalizeUnmockedModulePathPatterns(oldOptions, key);
         break;
       case 'haste':
-        value = {...options[key]};
+        value = {...oldOptions[key]};
         if (value.hasteImplModulePath != null) {
-          value.hasteImplModulePath = resolve(newOptions.resolver, {
+          const resolvedHasteImpl = resolve(newOptions.resolver, {
             filePath: replaceRootDirInPath(
               options.rootDir,
               value.hasteImplModulePath,
@@ -609,10 +654,12 @@ export default function normalize(
             key: 'haste.hasteImplModulePath',
             rootDir: options.rootDir,
           });
+
+          value.hasteImplModulePath = resolvedHasteImpl || undefined;
         }
         break;
       case 'projects':
-        value = (options[key] || [])
+        value = (oldOptions[key] || [])
           .map(project =>
             typeof project === 'string'
               ? _replaceRootDirTags(options.rootDir, project)
@@ -632,7 +679,7 @@ export default function normalize(
         {
           const replacedRootDirTags = _replaceRootDirTags(
             escapeGlobCharacters(options.rootDir),
-            options[key],
+            oldOptions[key],
           );
 
           if (replacedRootDirTags) {
@@ -645,12 +692,17 @@ export default function normalize(
         }
         break;
       case 'testRegex':
-        value = options[key]
-          ? [].concat(options[key]).map(replacePathSepForRegex)
-          : [];
+        {
+          const option = oldOptions[key];
+          value = option
+            ? (Array.isArray(option) ? option : [option]).map(
+                replacePathSepForRegex,
+              )
+            : [];
+        }
         break;
       case 'moduleFileExtensions': {
-        value = options[key];
+        value = oldOptions[key];
 
         if (
           Array.isArray(value) && // If it's the wrong type, it can throw at a later time
@@ -680,16 +732,17 @@ export default function normalize(
         break;
       }
       case 'bail': {
-        if (typeof options[key] === 'boolean') {
-          value = options[key] ? 1 : 0;
-        } else if (typeof options[key] === 'string') {
+        const bail = oldOptions[key];
+        if (typeof bail === 'boolean') {
+          value = bail ? 1 : 0;
+        } else if (typeof bail === 'string') {
           value = 1;
           // If Jest is invoked as `jest --bail someTestPattern` then need to
           // move the pattern from the `bail` configuration and into `argv._`
           // to be processed as an extra parameter
-          argv._.push(options[key]);
+          argv._.push(bail);
         } else {
-          value = options[key];
+          value = oldOptions[key];
         }
         break;
       }
@@ -746,10 +799,10 @@ export default function normalize(
       case 'watch':
       case 'watchAll':
       case 'watchman':
-        value = options[key];
+        value = oldOptions[key];
         break;
       case 'watchPlugins':
-        value = (options[key] || []).map(watchPlugin => {
+        value = (oldOptions[key] || []).map(watchPlugin => {
           if (typeof watchPlugin === 'string') {
             return {
               config: {},
@@ -770,7 +823,7 @@ export default function normalize(
         });
         break;
     }
-    // $FlowFixMe - automock is missing in GlobalConfig, so what
+    // @ts-ignore: automock is missing in GlobalConfig, so what
     newOptions[key] = value;
     return newOptions;
   }, newOptions);
@@ -779,16 +832,17 @@ export default function normalize(
   newOptions.testPathPattern = buildTestPathPattern(argv);
   newOptions.json = argv.json;
 
-  newOptions.testFailureExitCode = parseInt(newOptions.testFailureExitCode, 10);
+  newOptions.testFailureExitCode = parseInt(
+    (newOptions.testFailureExitCode as unknown) as string,
+    10,
+  );
 
-  for (const key of [
-    'lastCommit',
-    'changedFilesWithAncestor',
-    'changedSince',
-  ]) {
-    if (newOptions[key]) {
-      newOptions.onlyChanged = true;
-    }
+  if (
+    newOptions.lastCommit ||
+    newOptions.changedFilesWithAncestor ||
+    newOptions.changedSince
+  ) {
+    newOptions.onlyChanged = true;
   }
 
   if (argv.all) {
@@ -806,17 +860,20 @@ export default function normalize(
       ? 'all'
       : 'new';
 
-  newOptions.maxConcurrency = parseInt(newOptions.maxConcurrency, 10);
+  newOptions.maxConcurrency = parseInt(
+    (newOptions.maxConcurrency as unknown) as string,
+    10,
+  );
   newOptions.maxWorkers = getMaxWorkers(argv);
 
-  if (newOptions.testRegex.length && options.testMatch) {
+  if (newOptions.testRegex!.length && options.testMatch) {
     throw createConfigError(
       `  Configuration options ${chalk.bold('testMatch')} and` +
         ` ${chalk.bold('testRegex')} cannot be used together.`,
     );
   }
 
-  if (newOptions.testRegex.length && !options.testMatch) {
+  if (newOptions.testRegex!.length && !options.testMatch) {
     // Prevent the default testMatch conflicting with any explicitly
     // configured `testRegex` value
     newOptions.testMatch = [];
@@ -849,7 +906,7 @@ export default function normalize(
         if (
           !micromatch.some(
             replacePathSepForGlob(path.relative(options.rootDir, filename)),
-            newOptions.collectCoverageFrom,
+            newOptions.collectCoverageFrom!,
           )
         ) {
           return patterns;