+++ packages/jest-config/src/setFromArgv.ts
@@ -3,24 +3,21 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-import type {InitialOptions} from 'types/Config';
-import type {Argv} from 'types/Argv';
+import {Config} from '@jest/types';
 
 const specialArgs = ['_', '$0', 'h', 'help', 'config'];
 import {isJSONString} from './utils';
 
 export default function setFromArgv(
-  options: InitialOptions,
-  argv: Argv,
-): InitialOptions {
+  options: Config.InitialOptions,
+  argv: Config.Argv,
+): Config.InitialOptions {
   // $FlowFixMe: Seems like flow doesn't approve of string values
   const argvToOptions = Object.keys(argv)
     .filter(key => argv[key] !== undefined && specialArgs.indexOf(key) === -1)
-    .reduce((options: {[key: string]: mixed}, key) => {
+    .reduce((options: {[key: string]: unknown}, key) => {
       switch (key) {
         case 'coverage':
           options.collectCoverage = argv[key];