+++ packages/jest-config/src/validatePattern.ts
@@ -3,11 +3,9 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-export default function validatePattern(pattern: string) {
+export default function validatePattern(pattern?: string) {
   if (pattern) {
     try {
       // eslint-disable-next-line no-new